Make the workspace cycler animation parameters editable via chrome://gesture
Bug=174123
Test=Manual

Repro steps:
1) Run chrome with --ash-enable-workspace-scrubbing
2) Go to chrome://gesture
3) Modify the value of "Duration of the animations to start cycling"
4) Ensure that you have several maximized workspaces
5) Move fingers vertically on track pad. Notice difference in the duration of the animations to start cycling through workspaces
Repeat 3 - 5 as many times as you see fit

Review URL: https://chromiumcodereview.appspot.com/12212040

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@182243 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a3eeccb96cd74fe1fa73ab0be8487ffed0a18735

+++ ash/wm/workspace/workspace_cycler_animator.h
@@ -75,7 +75,7 @@ class ASH_EXPORT WorkspaceCyclerAnimator :
   void AbortAnimations();
 
   // Animate cycling by |scroll_delta|.
-  void AnimateCyclingByScrollDelta(float scroll_delta);
+  void AnimateCyclingByScrollDelta(double scroll_delta);
 
   // Returns the workspace which should be activated if the user does not do
   // any more cycling.
